requirements.txt
fastapi==0.85.0
Pillow==9.2.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==5.4.1
slowapi==0.1.7
torch==1.8.1
torchvision==0.9.1
uvicorn==0.18.3
vietocr==0.3.2